requirements.txt
# packages in environment at /home/zhouyuhua/anaconda3/envs/pytorch_env:
#
# Name                    Version                   Build  Channel
_libgcc_mutex             0.1                        main  
absl-py                   1.0.0                     <pip>
ca-certificates           2021.10.26           h06a4308_2  
cachetools                4.2.4                     <pip>
certifi                   2021.10.8        py37h06a4308_0  
charset-normalizer        2.0.9                     <pip>
cycler                    0.11.0                    <pip>
fonttools                 4.28.3                    <pip>
google-auth               2.3.3                     <pip>
google-auth-oauthlib      0.4.6                     <pip>
grpcio                    1.42.0                    <pip>
idna                      3.3                       <pip>
importlib-metadata        4.8.2                     <pip>
joblib                    1.1.0                     <pip>
kaggle                    1.5.12                    <pip>
kiwisolver                1.3.2                     <pip>
ld_impl_linux-64          2.35.1               h7274673_9  
libffi                    3.3                  he6710b0_2  
libgcc-ng                 9.1.0                hdf63c60_0  
libstdcxx-ng              9.1.0                hdf63c60_0  
Markdown                  3.3.6                     <pip>
matplotlib                3.5.0                     <pip>
mlxtend                   0.19.0                    <pip>
ncurses                   6.3                  h7f8727e_2  
netron                    5.4.1                     <pip>
numpy                     1.21.4                    <pip>
oauthlib                  3.1.1                     <pip>
opencv-python             4.5.4.60                  <pip>
openssl                   1.1.1l               h7f8727e_0  
packaging                 21.3                      <pip>
pandas                    1.3.4                     <pip>
Pillow                    8.4.0                     <pip>
pip                       21.2.2           py37h06a4308_0  
protobuf                  3.19.1                    <pip>
pyasn1                    0.4.8                     <pip>
pyasn1-modules            0.2.8                     <pip>
pyparsing                 3.0.6                     <pip>
python                    3.7.11               h12debd9_0  
python-dateutil           2.8.2                     <pip>
python-slugify            5.0.2                     <pip>
pytz                      2021.3                    <pip>
readline                  8.1                  h27cfd23_0  
requests                  2.26.0                    <pip>
requests-oauthlib         1.3.0                     <pip>
rsa                       4.8                       <pip>
scikit-learn              1.0.1                     <pip>
scipy                     1.7.3                     <pip>
seaborn                   0.11.2                    <pip>
setuptools                58.0.4           py37h06a4308_0  
setuptools-scm            6.3.2                     <pip>
six                       1.16.0                    <pip>
sqlite                    3.36.0               hc218d9a_0  
tb-nightly                2.8.0a20211209            <pip>
tensorboard               2.7.0                     <pip>
tensorboard-data-server   0.6.1                     <pip>
tensorboard-plugin-wit    1.8.0                     <pip>
text-unidecode            1.3                       <pip>
threadpoolctl             3.0.0                     <pip>
tk                        8.6.11               h1ccaba5_0  
tomli                     1.2.2                     <pip>
torch                     1.10.0                    <pip>
torchaudio                0.10.0                    <pip>
torchsummary              1.5.1                     <pip>
torchvision               0.11.1                    <pip>
tqdm                      4.62.3                    <pip>
typing_extensions         4.0.1                     <pip>
unet                      0.7.7                     <pip>
urllib3                   1.26.7                    <pip>
Werkzeug                  2.0.2                     <pip>
wheel                     0.37.0             pyhd3eb1b0_1  
xz                        5.2.5                h7b6447c_0  
zipp                      3.6.0                     <pip>
zlib                      1.2.11               h7b6447c_3  
